Basic Media Controls Working (in isolation)
Media controls are now working on the first 3 keys of layer 1. (The rest of the keyboard functionality is not working as the report descriptor is not properly configured and the rest of the applciation code needs to be changed to accomodate that.)

Key0 is mute
Key1 is vol up
Key2 is vol down.
Turns out the media report is acutally just a bitmask with the least significant 7 bits mapping to vol up, vol down, etc.
The current implementation is super dodgy but I wanted to commit this version before I try to properly integrate.