Accidental Key Spamming Fixed
Lines 929 and 930 were added to fix the following problem.

Sometime a letter would be sent to the PC but the key would not be "released" i.e. the computer would spam the letter. Seems like at some point the "release" packet wasn't being sent properly so I have a delay in between 2 sends of the "release" packet to make sure the keystroke is cleared.

Note the letters and media keys are working together in this commit but it is not the final implementation