Generic Report for Media and Keyboard working
I have implemented a generic report for the that is 9 bits long. This allows the report to be used for any kind of report as long as the USB drivers permit (and the report fits in the correct length). I have not removed the old report code yet.

With this GenericReport struct you can only report one report at a time. If you need to send multiple different keystrokes or reports then you instantiate and use multiple GenericReport structs.